--- src/users/fixtures/session.json
@@ -1,8 +1,0 @@
-{
-  "id": "session_01H5K05VP5CPCXJA5Z7G191GS4",
-  "created_at": "2023-07-18T04:40:14.954Z",
-  "expires_at": "2023-07-18T22:40:15.555Z",
-  "authorized_organizations": [],
-  "unauthorized_organizations": [],
-  "token": "really-long-token"
-}
