+++ src/users/serializers/authentication-response.serializer.ts
@@ -2,12 +2,10 @@ import {
   AuthenticationResponse,
   AuthenticationResponseResponse,
 } from '../interfaces';
-import { deserializeSession } from './session.serializer';
 import { deserializeUser } from './user.serializer';
 
 export const deserializeAuthenticationResponse = (
   authenticationResponse: AuthenticationResponseResponse,
 ): AuthenticationResponse => ({
-  session: deserializeSession(authenticationResponse.session),
   user: deserializeUser(authenticationResponse.user),
 });
