+++ src/users/serializers/index.ts
@@ -6,7 +6,7 @@ export * from './complete-password-reset-options.serializer';
 export * from './create-password-reset-challenge.serializer';
 export * from './create-user-options.serializer';
 export * from './send-magic-auth-code-options.serializer';
-export * from './session.serializer';
+export * from './user-authentication.serializer';
 export * from './update-user-options.serializer';
 export * from './update-user-password-options.serializer';
 export * from './user.serializer';
