+++ src/users/serializers/user-authentication.serializer.ts
@@ -1,22 +1,9 @@
 import {
-  Session,
-  SessionResponse,
   UnauthorizedOrganization,
   UnauthorizedOrganizationReason,
   UnauthorizedOrganizationReasonResponse,
   UnauthorizedOrganizationResponse,
-} from '../interfaces/session.interface';
-
-export const deserializeSession = (session: SessionResponse): Session => ({
-  id: session.id,
-  createdAt: session.created_at,
-  expiresAt: session.expires_at,
-  authorizedOrganizations: session.authorized_organizations,
-  unauthorizedOrganizations: session.unauthorized_organizations.map(
-    deserializeUnauthorizedOrganization,
-  ),
-  token: session.token,
-});
+} from '../interfaces/user-authentication.interface';
 
 export const deserializeUnauthorizedOrganization = (
   unauthorizedOrganization: UnauthorizedOrganizationResponse,
